feat: adicionando botoes para encerrar e deletar um quiz, melhorando logica de quais quizzes aparecem nas listas, icones svg e tipagem
