+++ src/components/icons/IconTrash.vue
@@ -1,0 +1,5 @@
+<template>
+  <svg xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" width="20" height="20" viewBox="0,0,256,256">
+    <g fill="#ffffff" fill-rule="nonzero" stroke="none" stroke-width="1" stroke-linecap="butt" stroke-linejoin="miter" stroke-miterlimit="10" stroke-dasharray="" stroke-dashoffset="0" font-family="none" font-weight="none" font-size="none" text-anchor="none" style="mix-blend-mode: normal"><g transform="scale(3.55556,3.55556)"><path d="M32.5,9c-4.136,0 -7.5,3.364 -7.5,7.5v1.5h-8c-2.209,0 -4,1.791 -4,4c0,2.209 1.791,4 4,4h0.23242l1.43945,25.91602c0.252,4.533 4.00688,8.08398 8.54688,8.08398h17.5625c4.54,0 8.29487,-3.55098 8.54688,-8.08398l1.43945,-25.91602h0.23242c2.209,0 4,-1.791 4,-4c0,-2.209 -1.791,-4 -4,-4h-8v-1.5c0,-4.136 -3.364,-7.5 -7.5,-7.5zM32.5,16h7c0.275,0 0.5,0.224 0.5,0.5v1.5h-8v-1.5c0,-0.276 0.225,-0.5 0.5,-0.5zM36,28c1.104,0 2,0.896 2,2v17.92383c0,1.105 -0.896,2 -2,2c-1.104,0 -2,-0.896 -2,-2v-17.92383c0,-1.104 0.896,-2 2,-2zM27.39258,28.00195c1.067,-0.022 2.02936,0.82569 2.06836,1.92969l0.625,18c0.038,1.104 -0.82764,2.02741 -1.93164,2.06641c-0.023,0.001 -0.04536,0.00195 -0.06836,0.00195c-1.073,0 -1.96005,-0.85164 -1.99805,-1.93164l-0.625,-18c-0.038,-1.104 0.82569,-2.02741 1.92969,-2.06641zM44.60742,28.00195c1.104,0.038 1.96769,0.96241 1.92969,2.06641l-0.625,18c-0.038,1.08 -0.92605,1.93164 -1.99805,1.93164c-0.023,0 -0.04536,-0.00095 -0.06836,-0.00195c-1.104,-0.038 -1.96964,-0.96241 -1.93164,-2.06641l0.625,-18c0.038,-1.104 0.97936,-1.95169 2.06836,-1.92969z"></path></g></g>
+  </svg>
+</template>
